fix ui impl

+++ xenon/src/game.cpp
@@ -1,43 +1,51 @@
 #include <xenon/features/game.hpp>
 #include <xenon/utility/random.hpp>
+#include <kiero/kiero.h>
 
 void Game::EnableUpdate() {
 
-	auto previousTime = std::chrono::steady_clock::now();
-	m_pSystem->g_fStartPlayTime = previousTime.time_since_epoch().count();
-
-	if (m_pConfigs->m_bUseCustomUI) {
-		m_pUIService->Init();
+	if (m_pSystem->IsInternal()) {
+		kiero::bind(8, (void**)&m_pUIService->oPresent, UIService::hkPresentWrapper);
 	}
+	else {
 
-	while (m_pConfigs->m_bUseUpdate) {
-		auto currentTime = std::chrono::steady_clock::now();
-		m_pSystem->g_fDeltaTime = std::chrono::duration<float>(currentTime - previousTime).count();
-		previousTime = currentTime;
+		auto previousTime = std::chrono::steady_clock::now();
+		m_pSystem->g_fStartPlayTime = previousTime.time_since_epoch().count();
 
-		if (m_pConfigs->m_bUseCustomUI && UIService::m_bShowMenu) {
-			m_pUIService->Update();
+		if (m_pConfigs->m_bUseCustomUI) {
+			m_pUIService->Init();
 		}
 
-		if (GetAsyncKeyState(m_pConfigs->m_nToggleUIKey) & 1) {
-			m_pUIService->m_bShowMenu = !UIService::m_bShowMenu;
-			if (UIService::m_bShowMenu) m_pUIService->SetMenuOpen();
-			else m_pUIService->SetMenuClose();
+		while (m_pConfigs->m_bUseUpdate) {
+			auto currentTime = std::chrono::steady_clock::now();
+			m_pSystem->g_fDeltaTime = std::chrono::duration<float>(currentTime - previousTime).count();
+			previousTime = currentTime;
+
+			if (GetAsyncKeyState(VK_ESCAPE)) {
+				continue;
+			}
+
+			TriggerEvent("Update");
+
+			HandleShortcuts();
+
+			if (m_pConfigs->m_bUseCustomUI && UIService::m_bShowMenu) {
+				m_pUIService->Update();
+			}
+
+			Update();
+
+			std::this_thread::sleep_for(std::chrono::milliseconds(5));
 		}
 
-		TriggerEvent("Update");
-
-		if (GetAsyncKeyState(VK_ESCAPE)) {
-			continue;
+		if (m_pConfigs->m_bUseCustomUI) {
+			m_pUIService->Destroy();
 		}
-
-		std::this_thread::sleep_for(std::chrono::milliseconds(5));
 	}
 
-	if (m_pConfigs->m_bUseCustomUI) {
-		m_pUIService->Destroy();
-	}
 }
+
+#pragma region Game:Private
 
 void Game::Update() {
 
@@ -72,7 +80,15 @@ void Game::Update() {
 
 }
 
-#pragma region Game:Private
+void Game::HandleShortcuts() {
 
+	if (GetAsyncKeyState(m_pConfigs->m_nToggleUIKey) & 1) {
+		m_pUIService->m_bShowMenu = !UIService::m_bShowMenu;
+		if (UIService::m_bShowMenu) m_pUIService->SetMenuOpen();
+		else m_pUIService->SetMenuClose();
+	}
+
+
+}
 
 #pragma endregion
